# docs/swagger.yaml
basePath: /
definitions:
...
  model.ProductType:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
...
  prodtype.paging:
    properties:
      nodes:
        items:
          $ref: '#/definitions/model.ProductType'
        type: array
      total:
        type: integer
    type: object
  prodtype.productType:
    properties:
      name:
        type: string
    type: object
...
host: localhost:7720
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Data portal API
  version: "1.0"
paths:
...
  /products/{product_id}/type:
    get:
      description: Get all productTypes
      operationId: get-all-productTypes
      parameters:
      - description: limt per page
        in: query
        name: limit
        type: string
      - description: page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/prodtype.paging'
      summary: Show all productTypes
      tags:
      - Type
    post:
      description: Create product type
      operationId: add-type
      parameters:
      - description: Type object
        in: body
        name: Type
        required: true
        schema:
          $ref: '#/definitions/prodtype.productType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ProductType'
        "400":
          description: Bad Request
          schema:
            items:
              type: string
            type: array
      summary: Create product type
      tags:
      - Type
...
swagger: "2.0"